Add test case for Summary component
